rename `footballData` to `sportData` in files
