+++ dotcom-rendering/src/server/server.dev.ts
@@ -14,16 +14,16 @@ import {
 } from './handler.article.web';
 import { handleEditionsCrossword } from './handler.editionsCrossword';
 import {
-	handleCricketMatchPage,
-	handleFootballMatchListPage,
-	handleFootballTablesPage,
-} from './handler.footballDataPage.web';
-import {
 	handleFront,
 	handleFrontJson,
 	handleTagPage,
 	handleTagPageJson,
 } from './handler.front.web';
+import {
+	handleCricketMatchPage,
+	handleFootballMatchListPage,
+	handleFootballTablesPage,
+} from './handler.sportDataPage.web';
 
 /** article URLs contain a part that looks like “2022/nov/25” */
 const ARTICLE_URL = /(\/\d{4}\/[a-z]{3}\/\d{2}\/)/;
